requirements.txt
openai>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.7.0
